refactor(selection-popover): extract virtual-keyboard-dismissal factory
Splits selection-popover.svelte (557 lines) per #1179: the on-screen
keyboard dismissal heuristic (frame bookkeeping, the two-source
ownership latch, the forgotten-set) moves verbatim into
createVirtualKeyboardDismissal, a factory matching createAnchoredOverlay's
shape (called at component top level, owns its own $effect).

This heuristic was previously only reachable through a full popover
mount plus real visualViewport/navigator.virtualKeyboard events. Adds
virtual-keyboard-dismissal.test.ts (mounting a small fixture) that
asserts the dismiss callback fires on an unowned keyboard
open-then-close transition, and is swallowed when the composer itself
holds focus during the same transition.

Pure code motion; behavior and public API unchanged.
